Implement the fileprotos as a subsystem
Allow for adding handling of more protocols for retrieving files.

This whole portion of the fileclient is a mess, but stripping this part
out would have a huge benefit of allowing people to more easily expand
it without messing around in the internals, maybe we need a git
file protocol?

rebasing from 2018 is hard

black hole sun

remove the unused imports

black is good

fake out pylint

fix docs
